Update to redis4cats 1.4.0
And bump also all other dependency versions, especially Scala itself.

The new redis4cats version brings as usual new operations that needed to be covered. As the transactions support became more fine-grained, not all of it can be fully supported, but the operations that worked before are still covered, so no regression in a narrower sense.